fix: rename vault command to encryption
