+++ internal/encryption/get/get.go
